fixed entrypoint.py; moved docker related files to /docker
